Fix cast parameter on java 1.8

+++ src/test/java/com/eternitywall/http/TestHttp.java
@@ -16,8 +16,7 @@ public class TestHttp {
     public void test() throws Exception {
         Request request= new Request(new URL("http://httpbin.org/status/418"));
         Response call = request.call();
-        assertEquals(418,call.getStatus());
-
+        assertEquals((Integer)418,call.getStatus());
 
     }
